New Project
This project is different from others by including storing and using data. This process is achieved with the help of localStorage object. This code works at any PC or laptop. I think this helps beginners like me to use a real data storing method. There are still a lot of things to do with this and I am planning to improve it a lot in next few days.